requirements.txt
pystray>=0.19
Pillow>=10.0
pyautogui>=0.9.54
pyinstaller>=6.0

pywin32>=306; sys_platform == "win32"
pyobjc>=10.0; sys_platform == "darwin"
